implemented polygon drawing, markers colors, removed debug logs

+++ src/app/maps/map.component.ts
@@ -5,6 +5,7 @@ import {Polygon} from "../models/polygon";
 import {Point} from "../models/point";
 import {user} from "../session";
 import {AuthenticationService} from "../services/authentication.service";
+import {LatLngLiteral} from "@agm/core";
 declare let $:any;
 const ICON_SIZE = 40;
 
@@ -25,7 +26,10 @@ export class MapComponent {
   customClassName: string = null;
   mapStyle: any = [{"elementType":"geometry","stylers":[{"color":"#1d2c4d"}]},{"elementType":"labels.text.fill","stylers":[{"color":"#8ec3b9"}]},{"elementType":"labels.text.stroke","stylers":[{"color":"#1a3646"}]},{"featureType":"administrative.country","elementType":"geometry.stroke","stylers":[{"color":"#4b6878"}]},{"featureType":"administrative.land_parcel","stylers":[{"visibility":"off"}]},{"featureType":"administrative.land_parcel","elementType":"labels.text.fill","stylers":[{"color":"#64779e"}]},{"featureType":"administrative.neighborhood","stylers":[{"visibility":"off"}]},{"featureType":"administrative.province","elementType":"geometry.stroke","stylers":[{"color":"#4b6878"}]},{"featureType":"landscape.man_made","elementType":"geometry.stroke","stylers":[{"color":"#334e87"}]},{"featureType":"landscape.natural","elementType":"geometry","stylers":[{"color":"#023e58"}]},{"featureType":"poi","elementType":"geometry","stylers":[{"color":"#283d6a"}]},{"featureType":"poi","elementType":"labels.text","stylers":[{"visibility":"off"}]},{"featureType":"poi","elementType":"labels.text.fill","stylers":[{"color":"#6f9ba5"}]},{"featureType":"poi","elementType":"labels.text.stroke","stylers":[{"color":"#1d2c4d"}]},{"featureType":"poi.park","elementType":"geometry.fill","stylers":[{"color":"#023e58"}]},{"featureType":"poi.park","elementType":"labels.text.fill","stylers":[{"color":"#3C7680"}]},{"featureType":"road","elementType":"geometry","stylers":[{"color":"#304a7d"}]},{"featureType":"road","elementType":"labels","stylers":[{"visibility":"off"}]},{"featureType":"road","elementType":"labels.text.fill","stylers":[{"color":"#98a5be"}]},{"featureType":"road","elementType":"labels.text.stroke","stylers":[{"color":"#1d2c4d"}]},{"featureType":"road.arterial","stylers":[{"visibility":"off"}]},{"featureType":"road.highway","elementType":"geometry","stylers":[{"color":"#2c6675"}]},{"featureType":"road.highway","elementType":"geometry.stroke","stylers":[{"color":"#255763"}]},{"featureType":"road.highway","elementType":"labels","stylers":[{"visibility":"off"}]},{"featureType":"road.highway","elementType":"labels.text.fill","stylers":[{"color":"#b0d5ce"}]},{"featureType":"road.highway","elementType":"labels.text.stroke","stylers":[{"color":"#023e58"}]},{"featureType":"road.local","stylers":[{"visibility":"off"}]},{"featureType":"transit","elementType":"labels.text.fill","stylers":[{"color":"#98a5be"}]},{"featureType":"transit","elementType":"labels.text.stroke","stylers":[{"color":"#1d2c4d"}]},{"featureType":"transit.line","elementType":"geometry.fill","stylers":[{"color":"#283d6a"}]},{"featureType":"transit.station","elementType":"geometry","stylers":[{"color":"#3a4762"}]},{"featureType":"water","elementType":"geometry","stylers":[{"color":"#0e1626"}]},{"featureType":"water","elementType":"labels.text","stylers":[{"visibility":"off"}]},{"featureType":"water","elementType":"labels.text.fill","stylers":[{"color":"#4e6d70"}]}];
   areaColor: string = "green";
+  markerColor: string = "white";
   errorMessage: string;
+  tempPoints: Array<LatLngLiteral> = [];
+  isFABVisible: boolean = false;
 
   constructor(private application: ApplicationRef,
               private pointService: PointService,
@@ -54,7 +58,11 @@ export class MapComponent {
 
     //get areas
     this.polygonService.list().subscribe(
-      list => this.areas = list,
+      list => {
+        // TODO: Polygon.polygon default value on API is NULL remove when fixed
+        list.splice(0, 1);
+        this.areas = list;
+      },
       error =>  this.errorMessage = <any>error
     );
   }
@@ -71,7 +79,10 @@ export class MapComponent {
     if(this.customTool == "DELETE") {
       let index = this.areas.indexOf(area);
       this.areas.splice(index, 1);
-      this.polygonService.remove(area.id);
+
+      if(area.id) {
+        this.polygonService.remove(area.id);
+      }
     }
   }
 
@@ -81,24 +92,15 @@ export class MapComponent {
         case "DELETE":
           break;
         case "AREA":
-          let polygon = new Polygon();
-
-          polygon.polygon = [
-            {lat: event.coords.lat, lng: event.coords.lng},
-            {lat: event.coords.lat+0.1, lng: event.coords.lng+0.1},
-            {lat: event.coords.lat+0.1, lng: event.coords.lng-0.1}
-          ];
-          polygon.color = this.areaColor;
-
-          this.areas.push(polygon);
-          this.polygonService.create(polygon);
+          this.tempPoints.push({lat: event.coords.lat, lng: event.coords.lng});
+          console.debug(this.tempPoints);
           break;
         default:
           let canvas = document.createElement("canvas");
           canvas.width = ICON_SIZE;
           canvas.height = ICON_SIZE;
           let ctx = canvas.getContext("2d");
-          ctx.fillStyle = "#ffffff";
+          ctx.fillStyle = this.markerColor;
           ctx.font = ICON_SIZE + "px FontAwesome";
           ctx.textAlign = "center";
           ctx.textBaseline = "middle";
@@ -118,11 +120,10 @@ export class MapComponent {
   }
 
   markerDragEnd(marker: Point, event: any) {
-    this.pointService.update(marker);
-  }
-
-  polygonDragEnd(area: Polygon, event: any) {
-    this.polygonService.update(area);
+    let index = this.markers.indexOf(marker);
+    this.markers[index].lat = event.coords.lat;
+    this.markers[index].lng = event.coords.lng;
+    this.pointService.update(this.markers[index]);
   }
 
   onToolPick(unicode, className, custom = null): void {
@@ -133,11 +134,29 @@ export class MapComponent {
   }
 
   onColorPick(color): void {
-    this.areaColor = color;
+    if(this.customTool == "AREA") {
+      this.areaColor = color;
+    } else {
+      this.markerColor = color;
+    }
+  }
+
+  createPolygon(): void {
+    let polygon = new Polygon();
+    polygon.color = this.areaColor;
+    polygon.polygon = this.tempPoints;
+    this.areas.push(polygon);
+    this.tempPoints = [];
+    this.polygonService.create(polygon);
+    this.getMapData();
   }
 
   onResetTool(): void {
     if(this.isToolActive) {
+      if(this.customTool == "AREA" && this.tempPoints.length > 0) {
+        this.createPolygon();
+      }
+
       this.activeTool = null;
       this.isToolActive = false;
       this.customTool = null;
@@ -147,6 +166,8 @@ export class MapComponent {
   }
 
   toggleFAB(toggle): void {
-    toggle ? $('.fixed-action-btn').openFAB() : $('.fixed-action-btn').closeFAB();
+    this.isFABVisible = toggle;
+    let selector = '.fixed-action-btn';
+    toggle ? $(selector).openFAB() : $(selector).closeFAB();
   }
 }
